test: extract shared SOAP test servers into soap_harness_test.go
newSOAPTestServer moves out of ptz_test.go, where it had ended up despite
already being shared with imaging_test.go, and is joined by a fault-serving
counterpart for exercising the error return of a client method rather than
its response mapping.

The fault is carried on HTTP 500 deliberately: internal/soap.Client.Call
decides success from the HTTP status and from whether the body parses, and
never consults Body.Fault, so the same fault served with a 200 would
unmarshal into an all-zero response and be reported to the caller as
success.
